+++ test/integration/dev/testApp/src/config/rendering.js
@@ -17,6 +17,21 @@ export default {
     getAdditionalDocumentContent: () => (
       <span style={{ display: 'none' }}>Additional Content</span>
     ),
+    processCSS: (tags) => (
+      [
+        <link key="hello-css" name="hello-css" />,
+      ].concat(tags)
+    ),
+    processJS: (tags) => (
+      [
+        <script key="hello-js" name="hello-js" />,
+      ].concat(tags)
+    ),
+    processMeta: (tags) => (
+      [
+        <meta key="hello-js" name="hello-meta" />,
+      ].concat(tags)
+    ),
     App,
   },
   client: {