+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.2] - 2017-11-16
+### Added
+- Experimental - Added processCSS, processJS and processMeta rendering config options
+
+## [0.7.1] - 2017-11-16
+### Added
+- Experimental - Added SSL DB connection option
+
 ## [0.7.0] - 2017-11-12
 ### Added
 - Added the ability to add wrappers around client/server side App components